move jsonrpc header into gotracer
