+++ bpf/gotracer/protocol_jsonrpc.h
